requirements.txt
streamlit==1.28.1
openai==1.3.8
python-dotenv==1.0.0
pydantic==2.5.0